+++ src/main/java/es/in2/issuer/backend/shared/domain/model/port/IssuerProperties.java
@@ -8,7 +8,6 @@ public interface IssuerProperties {
     String getVerifierUrl();
     String getVerifierInternalUrl();
     String getDefaultLang();
-    String getAdminOrganizationId();
     String getSysTenant();
     String getManagementTokenOrgIdJsonPath();
     String getManagementTokenAdminPowerFunction();