+++ src/main/java/es/in2/issuer/backend/shared/infrastructure/config/properties/AppProperties.java
@@ -17,7 +17,6 @@ public record AppProperties(
         @URL String verifierInternalUrl,
         @NotBlank @URL String walletUrl,
         @NotBlank String defaultLang,
-        @NotBlank String adminOrganizationId,
         @NotBlank String sysTenant,
         ManagementToken managementToken
 ) {